refactor(spp): split proof_bundle.go grab-bag into focused files
Per architecture review. proof_bundle.go (1137 lines) mixed DTOs, witness
assembly, P256 helpers, hashing, and field/bigint utilities. Extracted:
- request.go    — JSON request/response DTOs.
- p256.go       — P256 witness construction helpers.
- fieldutil.go  — field/hex/bigint parsing and conversion utilities.
proof_bundle.go now holds the bundle entry points + witness assembly + the
external-data / signed-amount hashing (~709 lines). No behavior change; full
SPP suite + vet + gofmt green.
