+++ prover/server/prover/spp/proof_bundle.go
@@ -1,12 +1,8 @@
 package spp
 
 import (
-	"crypto/ecdsa"
-	"crypto/elliptic"
-	"crypto/rand"
 	"crypto/sha256"
 	"encoding/binary"
-	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"math/big"
@@ -16,154 +12,7 @@ import (
 	"light/light-prover/prover/common"
 
 	"github.com/consensys/gnark/frontend"
-	"github.com/consensys/gnark/std/math/emulated"
-	gnarkecdsa "github.com/consensys/gnark/std/signature/ecdsa"
 )
-
-type ProofBundleRequest struct {
-	SolanaSignerPubkey string                    `json:"solana_signer_pubkey"`
-	Transactions       []ProofTransactionRequest `json:"transactions"`
-}
-
-type ProofTransactionRequest struct {
-	Name                     string              `json:"name"`
-	InstructionDiscriminator uint8               `json:"instruction_discriminator"`
-	ExpiryUnixTs             uint64              `json:"expiry_unix_ts"`
-	SenderViewTag            string              `json:"sender_view_tag"`
-	RelayerFee               uint16              `json:"relayer_fee"`
-	PublicAmountMode         uint8               `json:"public_amount_mode"`
-	PublicSolAmount          *uint64             `json:"public_sol_amount"`
-	PublicSplAmount          *uint64             `json:"public_spl_amount"`
-	PublicSplAssetPubkey     string              `json:"public_spl_asset_pubkey"`
-	EncryptedUtxos           string              `json:"encrypted_utxos"`
-	UserSolAccount           string              `json:"user_sol_account"`
-	UserSplTokenAccount      string              `json:"user_spl_token_account"`
-	SplTokenInterface        string              `json:"spl_token_interface"`
-	StateEntries             []ProofStateEntry   `json:"state_entries"`
-	Inputs                   []ProofInputRequest `json:"inputs"`
-	Outputs                  []ProofUtxoRequest  `json:"outputs"`
-	UtxoTreeRootIndex        []uint16            `json:"utxo_tree_root_index"`
-	NullifierTreeRootIndex   []uint16            `json:"nullifier_tree_root_index"`
-	NullifierEntries         []string            `json:"nullifier_entries"`
-	ProgramIDHashChain       string              `json:"program_id_hashchain"`
-	P256SignerPubkey         string              `json:"p256_signer_pubkey,omitempty"`
-	P256SignatureR           string              `json:"p256_signature_r,omitempty"`
-	P256SignatureS           string              `json:"p256_signature_s,omitempty"`
-}
-
-type ProofStateEntry struct {
-	Index uint64 `json:"index"`
-	Hash  string `json:"hash"`
-}
-
-type ProofInputRequest struct {
-	Utxo      ProofUtxoRequest `json:"utxo"`
-	LeafIndex uint64           `json:"leaf_index"`
-	// NullifierSecret is the authoritative secret for this input. When set it
-	// takes precedence over Utxo.OwnerNullifierSecret, which is only a fallback
-	// for recomputing owner components when no input-level secret is supplied.
-	// See ownerComponents.
-	NullifierSecret string `json:"nullifier_secret"`
-}
-
-type ProofUtxoRequest struct {
-	Domain            string `json:"domain"`
-	Owner             string `json:"owner"`
-	OwnerSolanaPubkey string `json:"owner_solana_pubkey"`
-	OwnerP256Pubkey   string `json:"owner_p256_pubkey,omitempty"`
-	// OwnerNullifierSecret is a fallback used only when this UTXO has no
-	// enclosing ProofInputRequest.NullifierSecret (e.g. a bare output UTXO whose
-	// owner hash must be recomputed). For inputs, set NullifierSecret instead.
-	OwnerNullifierSecret string `json:"owner_nullifier_secret,omitempty"`
-	AssetID              string `json:"asset_id"`
-	AssetAmount          string `json:"asset_amount"`
-	Blinding             string `json:"blinding"`
-	DataHash             string `json:"data_hash"`
-	PolicyData           string `json:"policy_data"`
-	PolicyProgramID      string `json:"policy_program_id"`
-}
-
-type ProofBundle struct {
-	Shape                 Shape              `json:"shape"`
-	SolanaSignerPubkeyHex string             `json:"solana_signer_pubkey"`
-	Transactions          []ProofTransaction `json:"transactions"`
-}
-
-type ProofTransaction struct {
-	Name                    string              `json:"name"`
-	ExpiryUnixTs            uint64              `json:"expiry_unix_ts"`
-	SenderViewTag           string              `json:"sender_view_tag"`
-	Proof                   *common.Proof       `json:"proof"`
-	RelayerFee              uint16              `json:"relayer_fee"`
-	Nullifiers              []string            `json:"nullifiers"`
-	OutputUtxoHashes        []string            `json:"output_utxo_hashes"`
-	UtxoTreeRootIndex       []uint16            `json:"utxo_tree_root_index"`
-	NullifierTreeRootIndex  []uint16            `json:"nullifier_tree_root_index"`
-	PrivateTxHash           string              `json:"private_tx_hash"`
-	PublicAmountMode        uint8               `json:"public_amount_mode"`
-	PublicSolAmount         *uint64             `json:"public_sol_amount"`
-	PublicSplAmount         *uint64             `json:"public_spl_amount"`
-	PublicSplAssetPubkey    string              `json:"public_spl_asset_pubkey"`
-	EncryptedUtxos          string              `json:"encrypted_utxos"`
-	PublicInputHash         string              `json:"public_input_hash"`
-	ExternalDataHash        string              `json:"external_data_hash"`
-	UserSolAccount          string              `json:"user_sol_account"`
-	UserSplTokenAccount     string              `json:"user_spl_token_account"`
-	SplTokenInterface       string              `json:"spl_token_interface"`
-	InUtxoSignerIndices     []int               `json:"in_utxo_signer_indices"`
-	OutputUtxos             []ProofUtxoResponse `json:"output_utxos"`
-	DebugInputUtxoHashes    []string            `json:"debug_input_utxo_hashes"`
-	DebugOutputUtxoHashes   []string            `json:"debug_output_utxo_hashes"`
-	DebugUtxoTreeRoots      []string            `json:"debug_utxo_tree_roots"`
-	DebugNullifierTreeRoots []string            `json:"debug_nullifier_tree_roots"`
-}
-
-type ProofSigningPayloadBundle struct {
-	Shape                 Shape                            `json:"shape"`
-	SolanaSignerPubkeyHex string                           `json:"solana_signer_pubkey"`
-	Transactions          []ProofSigningPayloadTransaction `json:"transactions"`
-}
-
-type ProofSigningPayloadTransaction struct {
-	Name                  string `json:"name"`
-	PrivateTxHash         string `json:"private_tx_hash"`
-	RequiresP256Signature bool   `json:"requires_p256_signature"`
-}
-
-type ProofUtxoResponse struct {
-	Utxo ProofUtxoRequest `json:"utxo"`
-	Hash string           `json:"hash"`
-}
-
-type proofInput struct {
-	utxo            Utxo
-	utxoRequest     ProofUtxoRequest
-	leafIndex       uint64
-	nullifierSecret *big.Int
-	ownerKeyHash    *big.Int
-	nullifierPk     *big.Int
-	isP256          bool
-}
-
-type proofDebug struct {
-	inputHashes              []*big.Int
-	outputHashes             []*big.Int
-	nullifiers               []*big.Int
-	inUtxoSignerIndices      []int
-	requiresP256OwnerWitness bool
-}
-
-type proofExternalData struct {
-	InstructionDiscriminator uint8
-	SenderViewTag            [32]byte
-	RelayerFee               uint16
-	PublicSolAmount          uint64
-	PublicSplAmount          uint64
-	UserSolAccount           [32]byte
-	UserSplToken             [32]byte
-	SplTokenInterface        [32]byte
-	EncryptedUtxos           []byte
-}
 
 func WriteProofBundle(ps *ProofSystem, requestPath string, outputPath string) error {
 	bytes, err := os.ReadFile(requestPath)
@@ -809,106 +658,6 @@ func ownerComponents(input ProofUtxoRequest, inputNullifierSecret *big.Int) (own
 	return ownerKey{keyHash: keyHash, nullifierPk: nullifierPk, isP256: isP256}, nil
 }
 
-func p256WitnessForTransaction(
-	tx ProofTransactionRequest,
-	privateTxHash *big.Int,
-	requiresP256 bool,
-	allowMissingSignature bool,
-) (gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr], gnarkecdsa.Signature[emulated.P256Fr], error) {
-	msg := proofFieldBytes(privateTxHash)
-	if !requiresP256 && strings.TrimSpace(tx.P256SignerPubkey) == "" {
-		return dummyP256Witness(msg[:])
-	}
-	if allowMissingSignature && (strings.TrimSpace(tx.P256SignerPubkey) == "" || tx.P256SignatureR == "" || tx.P256SignatureS == "") {
-		return dummyP256Witness(msg[:])
-	}
-	pub, err := p256PubkeyWitness(tx.P256SignerPubkey)
-	if err != nil {
-		return gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr]{}, gnarkecdsa.Signature[emulated.P256Fr]{}, fmt.Errorf("p256_signer_pubkey: %w", err)
-	}
-	if tx.P256SignatureR == "" || tx.P256SignatureS == "" {
-		if requiresP256 {
-			return gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr]{}, gnarkecdsa.Signature[emulated.P256Fr]{}, fmt.Errorf("p256_signature_r and p256_signature_s are required for P256 inputs")
-		}
-		return dummyP256Witness(msg[:])
-	}
-	r, err := parseP256Scalar(tx.P256SignatureR)
-	if err != nil {
-		return gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr]{}, gnarkecdsa.Signature[emulated.P256Fr]{}, fmt.Errorf("p256_signature_r: %w", err)
-	}
-	s, err := parseP256Scalar(tx.P256SignatureS)
-	if err != nil {
-		return gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr]{}, gnarkecdsa.Signature[emulated.P256Fr]{}, fmt.Errorf("p256_signature_s: %w", err)
-	}
-	return pub, gnarkecdsa.Signature[emulated.P256Fr]{
-		R: emulated.ValueOf[emulated.P256Fr](r),
-		S: emulated.ValueOf[emulated.P256Fr](s),
-	}, nil
-}
-
-func p256PubkeyWitness(compressedHex string) (gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr], error) {
-	compressed, err := parseHexBytes(compressedHex)
-	if err != nil {
-		return gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr]{}, err
-	}
-	if len(compressed) != 33 {
-		return gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr]{}, fmt.Errorf("expected 33-byte compressed P256 public key, got %d", len(compressed))
-	}
-	x, y := elliptic.UnmarshalCompressed(elliptic.P256(), compressed)
-	if x == nil || y == nil {
-		return gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr]{}, fmt.Errorf("invalid compressed P256 public key")
-	}
-	return gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr]{
-		X: emulated.ValueOf[emulated.P256Fp](x),
-		Y: emulated.ValueOf[emulated.P256Fp](y),
-	}, nil
-}
-
-func dummyP256Witness(msg []byte) (gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr], gnarkecdsa.Signature[emulated.P256Fr], error) {
-	priv, err := fixedP256PrivateKey(big.NewInt(7))
-	if err != nil {
-		return gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr]{}, gnarkecdsa.Signature[emulated.P256Fr]{}, err
-	}
-	r, s, err := ecdsa.Sign(rand.Reader, priv, msg)
-	if err != nil {
-		return gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr]{}, gnarkecdsa.Signature[emulated.P256Fr]{}, err
-	}
-	return gnarkecdsa.PublicKey[emulated.P256Fp, emulated.P256Fr]{
-			X: emulated.ValueOf[emulated.P256Fp](priv.PublicKey.X),
-			Y: emulated.ValueOf[emulated.P256Fp](priv.PublicKey.Y),
-		}, gnarkecdsa.Signature[emulated.P256Fr]{
-			R: emulated.ValueOf[emulated.P256Fr](r),
-			S: emulated.ValueOf[emulated.P256Fr](s),
-		}, nil
-}
-
-func fixedP256PrivateKey(d *big.Int) (*ecdsa.PrivateKey, error) {
-	curve := elliptic.P256()
-	if d.Sign() <= 0 || d.Cmp(curve.Params().N) >= 0 {
-		return nil, fmt.Errorf("invalid P256 private scalar")
-	}
-	x, y := curve.ScalarBaseMult(d.Bytes())
-	return &ecdsa.PrivateKey{
-		PublicKey: ecdsa.PublicKey{
-			Curve: curve,
-			X:     x,
-			Y:     y,
-		},
-		D: d,
-	}, nil
-}
-
-func parseP256Scalar(value string) (*big.Int, error) {
-	parsed, err := parseBigInt(value)
-	if err != nil {
-		return nil, err
-	}
-	if parsed.Sign() <= 0 || parsed.Cmp(elliptic.P256().Params().N) >= 0 {
-		return nil, fmt.Errorf("scalar is outside P256 scalar field")
-	}
-	return parsed, nil
-}
-
 func proofExternalDataFieldHash(data proofExternalData) *big.Int {
 	hasher := sha256.New()
 	hasher.Write([]byte{data.InstructionDiscriminator})
@@ -957,184 +706,4 @@ func signedSplAmount(mode uint8, amount uint64) *big.Int {
 	default:
 		return new(big.Int).SetUint64(amount)
 	}
-}
-
-func optionalU64(value *uint64) uint64 {
-	if value == nil {
-		return 0
-	}
-	return *value
-}
-
-func optionalField(value string) (*big.Int, error) {
-	if value == "" {
-		return big.NewInt(0), nil
-	}
-	return parseField(value)
-}
-
-func parseField(value string) (*big.Int, error) {
-	out, err := parseBigInt(value)
-	if err != nil {
-		return nil, err
-	}
-	if err := validateFieldElement("field", out); err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func parseBigInt(value string) (*big.Int, error) {
-	value = strings.TrimSpace(value)
-	value = strings.TrimPrefix(value, "0x")
-	if value == "" {
-		return nil, fmt.Errorf("empty field")
-	}
-	base := 10
-	if len(value) > 20 || strings.IndexFunc(value, func(r rune) bool {
-		return (r >= 'a' && r <= 'f') || (r >= 'A' && r <= 'F')
-	}) >= 0 {
-		base = 16
-	}
-	out, ok := new(big.Int).SetString(value, base)
-	if !ok {
-		return nil, fmt.Errorf("invalid field %q", value)
-	}
-	return out, nil
-}
-
-func parseHex32(value string) ([32]byte, error) {
-	bytes, err := parseHexBytes(value)
-	if err != nil {
-		return [32]byte{}, err
-	}
-	if len(bytes) != 32 {
-		return [32]byte{}, fmt.Errorf("expected 32 bytes, got %d", len(bytes))
-	}
-	var out [32]byte
-	copy(out[:], bytes)
-	return out, nil
-}
-
-func parseOptionalHex32(value string) ([32]byte, error) {
-	if strings.TrimSpace(value) == "" {
-		return [32]byte{}, nil
-	}
-	return parseHex32(value)
-}
-
-func parseHexBytes(value string) ([]byte, error) {
-	value = strings.TrimSpace(strings.TrimPrefix(value, "0x"))
-	if value == "" {
-		return nil, nil
-	}
-	out, err := hex.DecodeString(value)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func proofZeroUtxo() Utxo {
-	return Utxo{
-		Domain:          big.NewInt(0),
-		Owner:           big.NewInt(0),
-		AssetID:         big.NewInt(0),
-		AssetAmount:     big.NewInt(0),
-		Blinding:        big.NewInt(0),
-		DataHash:        big.NewInt(0),
-		PolicyData:      big.NewInt(0),
-		PolicyProgramID: big.NewInt(0),
-	}
-}
-
-func toProofCircuitFields(utxo Utxo) UtxoCircuitFields {
-	return UtxoCircuitFields{
-		Domain:          utxo.Domain,
-		Owner:           utxo.Owner,
-		AssetID:         utxo.AssetID,
-		AssetAmount:     utxo.AssetAmount,
-		Blinding:        utxo.Blinding,
-		DataHash:        utxo.DataHash,
-		PolicyData:      utxo.PolicyData,
-		PolicyProgramID: utxo.PolicyProgramID,
-	}
-}
-
-func proofZeroVariableSlice(n int) []frontend.Variable {
-	out := make([]frontend.Variable, n)
-	for i := range out {
-		out[i] = big.NewInt(0)
-	}
-	return out
-}
-
-func fillProofPath(path []frontend.Variable, dirs []frontend.Variable, siblings []*big.Int, directions []int) {
-	for i := range siblings {
-		path[i] = siblings[i]
-		dirs[i] = big.NewInt(int64(directions[i]))
-	}
-}
-
-func proofBigIntsToVariables(values []*big.Int) []frontend.Variable {
-	out := make([]frontend.Variable, len(values))
-	for i, value := range values {
-		out[i] = value
-	}
-	return out
-}
-
-func proofVariablesToBigInts(values []frontend.Variable) ([]*big.Int, error) {
-	out := make([]*big.Int, len(values))
-	for i, value := range values {
-		switch v := value.(type) {
-		case *big.Int:
-			out[i] = new(big.Int).Set(v)
-		case int:
-			out[i] = big.NewInt(int64(v))
-		case int64:
-			out[i] = big.NewInt(v)
-		default:
-			return nil, fmt.Errorf("spp: unexpected witness variable type %T", value)
-		}
-	}
-	return out, nil
-}
-
-func proofTrimTrailingZeroHexes(values []*big.Int) []string {
-	end := len(values)
-	for end > 0 && values[end-1].Sign() == 0 {
-		end--
-	}
-	out := make([]string, end)
-	for i := 0; i < end; i++ {
-		out[i] = proofFieldHex(values[i])
-	}
-	return out
-}
-
-func proofBigIntHexes(values []*big.Int) []string {
-	out := make([]string, len(values))
-	for i, value := range values {
-		out[i] = proofFieldHex(value)
-	}
-	return out
-}
-
-func proofFieldHex(value *big.Int) string {
-	return fmt.Sprintf("%064x", value)
-}
-
-func proofBytesHex(value []byte) string {
-	return fmt.Sprintf("%x", value)
-}
-
-func proofFieldBytes(value *big.Int) [32]byte {
-	var out [32]byte
-	if value == nil {
-		return out
-	}
-	bytes := value.Bytes()
-	copy(out[32-len(bytes):], bytes)
-	return out
 }
